Search and Calculate Agentic AI/requirements.txt
langchain>=0.2.0
google-generativeai>=0.5.0
duckduckgo-search>=5.3.0
newspaper3k>=0.2.8
beautifulsoup4>=4.12.3
requests>=2.31.0
sentence-transformers>=2.6.1
faiss-cpu>=1.8.0
